pyproject.toml
[tool.poetry]
name = "blog-test"
version = "0.1.0"
description = ""
authors = ["antony_kosenko <kosenko.anton.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.13"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
django-resized = "^1.0.2"
djangochannelsrestframework = "^1.2.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
django-mptt = "^0.16.0"
requests = "^2.31.0"
django-debug-toolbar = "^4.3.0"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
django-mysql = "^4.13.0"
mysqlclient = "2.1.1"
django-filter = "^24.2"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2024.1"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
